requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.6.1
dnspython==2.4.2
effdet==0.4.1
emoji==2.8.0
et-xmlfile==1.1.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.327
langdetect==1.0.9
langsmith==0.0.56
layoutparser==0.3.4
loguru==0.7.2
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
mdurl==0.1.2
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
olefile==0.46
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==0.28.1
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pinecone-client==2.2.4
portalocker==2.8.2
protobuf==4.24.4
pyarrow==14.0.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pypandoc==1.12
pyparsing==3.1.1
pypdfium2==4.23.1
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.0.1
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
rapidfuzz==3.5.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
streamlit==1.28.0
streamlit-chat==0.0.2.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
timm==0.9.8
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.34.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.10.28
unstructured-inference==0.7.10
unstructured.pytesseract==0.3.12
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
win32-setctime==1.1.0
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.2
zipp==3.17.0
